backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django_cors_headers==3.13.0 
django-environ==0.9.0
djangorestframework-simplejwt==5.2.2